// src/app/api/models/project.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Company } from '../models/company';
import { ProjectAddress } from '../models/project-address';
export interface Project {
  address?: ProjectAddress;
  company?: Company;
  companyId?: string;
  createdAt?: string;
  description?: string | null;
  id?: string;
  isActive?: boolean;
  name?: string | null;
  updatedAt?: string;
}