// src/app/api/models/project-address-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProjectAddressDto {
  city?: string | null;
  country?: string | null;
  houseNumber?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  postalCode?: string | null;
  street?: string | null;
}